Build of version 1.0.1
Build of version 1.0.1 which fixes some problems caused by differences in the game files between the steam and the itch.io version. The settings menu has also change. In addition a small bug that prevented you from using the gauss gun when running a TAS (even when the TAS had finished).